Remove external dependency.
