+++ chain.go
@@ -5,8 +5,6 @@ import (
 	"runtime"
 	"sync"
 	"time"
-
-	"github.com/mixer/clock"
 )
 
 // JobWrapper decorates the given Job with some behavior.
@@ -61,12 +59,12 @@ func Recover(logger Logger) JobWrapper {
 // previous one is complete. Jobs running after a delay of more than a minute
 // have the delay logged at Info.
 func DelayIfStillRunning(logger Logger) JobWrapper {
-	return DelayIfStillRunningWithClock(logger, clock.C)
+	return DelayIfStillRunningWithClock(logger, defaultClock{})
 }
 
 // DelayIfStillRunningWithClock behaves identically to DelayIfStillRunning but
 // uses the provided Clock for measuring the delay, for use in testing.
-func DelayIfStillRunningWithClock(logger Logger, clk clock.Clock) JobWrapper {
+func DelayIfStillRunningWithClock(logger Logger, clk Clock) JobWrapper {
 	return func(j Job) Job {
 		var mu sync.Mutex
 		return FuncJob(func() {
