+++ go.mod
@@ -1,5 +1,3 @@
 module github.com/robfig/cron/v3
 
 go 1.12
-
-require github.com/mixer/clock v0.0.0-20190507173039-c311c17adb1f
